Adjust efficiency curve for pump setting
Assumption is that model efficiency curve is correct for setting =
N2/N1 = 1.0, and that efficiency is “unaffected by speed” — which is a
bad way of saying that you use the efficiency at Q1 even though the
correct flow is Q2 = (N2/N1)*Q1. These appear to be standard
assumptions for design purposes that go along with the application of
affinity laws.